clean up warnings

+++ schedulers/mt/bench/bm_copy.cpp
@@ -22,8 +22,8 @@ using namespace gr;
 int main(int argc, char* argv[])
 {
     uint64_t samples;
-    int nblocks;
-    int nthreads;
+    unsigned int nblocks;
+    unsigned int nthreads;
     int veclen;
     int buffer_type;
     int buffer_size;
@@ -37,9 +37,9 @@ int main(int argc, char* argv[])
         po::value<uint64_t>(&samples)->default_value(15000000),
         "Number of samples")(
         "veclen", po::value<int>(&veclen)->default_value(1), "Vector Length")(
-        "nblocks", po::value<int>(&nblocks)->default_value(1), "Number of copy blocks")(
+        "nblocks", po::value<unsigned int>(&nblocks)->default_value(1), "Number of copy blocks")(
         "nthreads",
-        po::value<int>(&nthreads)->default_value(0),
+        po::value<unsigned int>(&nthreads)->default_value(0),
         "Number of threads (0: tpb")(
         "buffer",
         po::value<int>(&buffer_type)->default_value(1),
@@ -77,7 +77,7 @@ int main(int argc, char* argv[])
         auto head = blocks::head::make(sizeof(gr_complex) * veclen, samples / veclen);
         auto snk = blocks::null_sink::make(sizeof(gr_complex) * veclen);
         std::vector<blocks::copy::sptr> copy_blks(nblocks);
-        for (int i = 0; i < nblocks; i++) {
+        for (unsigned int i = 0; i < nblocks; i++) {
             copy_blks[i] = blocks::copy::make(sizeof(gr_complex) * veclen);
         }
         flowgraph_sptr fg(new flowgraph());
@@ -85,7 +85,7 @@ int main(int argc, char* argv[])
         if (buffer_type == 0) {
             fg->connect(src, 0, head, 0);
             fg->connect(head, 0, copy_blks[0], 0);
-            for (int i = 0; i < nblocks - 1; i++) {
+            for (unsigned int i = 0; i < nblocks - 1; i++) {
                 fg->connect(copy_blks[i], 0, copy_blks[i + 1], 0);
             }
             fg->connect(copy_blks[nblocks - 1], 0, snk, 0);
@@ -93,7 +93,7 @@ int main(int argc, char* argv[])
         } else {
             fg->connect(src, 0, head, 0)->set_custom_buffer(VMCIRC_BUFFER_ARGS);
             fg->connect(head, 0, copy_blks[0], 0)->set_custom_buffer(VMCIRC_BUFFER_ARGS);
-            for (int i = 0; i < nblocks - 1; i++) {
+            for (unsigned int i = 0; i < nblocks - 1; i++) {
                 fg->connect(copy_blks[i], 0, copy_blks[i + 1], 0)->set_custom_buffer(VMCIRC_BUFFER_ARGS);
             }
             fg->connect(copy_blks[nblocks - 1], 0, snk, 0)->set_custom_buffer(VMCIRC_BUFFER_ARGS);
@@ -110,7 +110,7 @@ int main(int argc, char* argv[])
         if (nthreads > 0) {
             int blks_per_thread = nblocks / nthreads;
 
-            for (int i = 0; i < nthreads; i++) {
+            for (unsigned int i = 0; i < nthreads; i++) {
                 std::vector<block_sptr> block_group;
                 if (i == 0) {
                     block_group.push_back(src);
@@ -121,8 +121,8 @@ int main(int argc, char* argv[])
                     block_group.push_back(copy_blks[i * blks_per_thread + j]);
                 }
 
-                if (i == static_cast<unsigned int>(nthreads - 1)) {
-                    for (int j = 0; j < (nblocks - nthreads * blks_per_thread); j++) {
+                if (i == nthreads - 1) {
+                    for (unsigned int j = 0; j < (nblocks - nthreads * blks_per_thread); j++) {
                         block_group.push_back(copy_blks[(i + 1) * blks_per_thread + j]);
                     }
                     block_group.push_back(snk);
